Move example data structure to SamplesUtils
